Fine minishell

+++ 07-Minishell/libft/ft_lstsize.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_lstsize.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: dripanuc <dripanuc@student.42.fr>          +#+  +:+       +#+        */
+/*   By: mpatrini <mpatrini@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/01/13 22:06:05 by dripanuc          #+#    #+#             */
-/*   Updated: 2022/01/14 17:09:34 by dripanuc         ###   ########.fr       */
+/*   Created: 2022/01/12 06:57:05 by mpatrini          #+#    #+#             */
+/*   Updated: 2022/01/12 07:03:45 by mpatrini         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_lstsize(t_list *lst)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (lst)
